Fix compiler warning

+++ ATmega128 C Project/navigation.c
@@ -100,7 +100,8 @@ static void nfCalcPosition(RobotGlobalData *sys)
 	char str[100]; // serial output string
 	sprintf(str, "POS,%6.3f,%6.3f,%6.1f,%1d,%1d,%4d,%4d,%4i\r\n",
 		sys->pos.x, sys->pos.y, sys->pos.heading * 180. / M_PI,	sys->pos.leftPulses,
-		sys->pos.rightPulses, sys->pos.leftTotal, sys->pos.rightTotal, sys->timeStamp);
+		sys->pos.rightPulses, sys->pos.leftTotal, sys->pos.rightTotal, (unsigned int)
+		sys->timeStamp);
 	
 	uartOutputString(str);	
 }
